# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Golf Pool Scoreboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Tournament {
  id: string;
  name: string;
  year: number;
  /** @nullable */
  espnEventId?: string | null;
  status: string;
  currentRound: number;
  isActive: boolean;
  createdAt: string;
}

export interface PoolMember {
  id: string;
  name: string;
  createdAt: string;
}

export interface GolferRoundDetail {
  golferId: string;
  golferName: string;
  /** @nullable */
  scoreToPar?: number | null;
  holesCompleted: number;
  isCut: boolean;
  isWd: boolean;
  isDq: boolean;
  isPenalty: boolean;
  /** @nullable */
  teeTime?: string | null;
  /** @nullable */
  counted?: boolean | null;
}

export interface RoundScore {
  roundNumber: number;
  /** @nullable */
  score?: number | null;
  golferDetails: GolferRoundDetail[];
}

export interface LeaderboardEntry {
  rank: number;
  poolMemberId: string;
  name: string;
  /** @nullable */
  toPar?: number | null;
  thru: string;
  /** @nullable */
  today?: number | null;
  /** @nullable */
  r1?: number | null;
  /** @nullable */
  r2?: number | null;
  /** @nullable */
  r3?: number | null;
  /** @nullable */
  r4?: number | null;
  rounds: RoundScore[];
}

export interface ScoreboardResponse {
  tournament: Tournament;
  /** @nullable */
  lastUpdated?: string | null;
  /** @nullable */
  nextUpdate?: string | null;
  refreshIntervalMinutes: number;
  leaderboard: LeaderboardEntry[];
}

export interface ManualScore {
  id: string;
  tournamentId: string;
  poolMemberId: string;
  poolMemberName: string;
  /** @nullable */
  r1?: number | null;
  /** @nullable */
  r2?: number | null;
  /** @nullable */
  r3?: number | null;
  /** @nullable */
  r4?: number | null;
  /** @nullable */
  toPar?: number | null;
  /** @nullable */
  updatedBy?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ManualScoreboardResponse {
  tournament: Tournament;
  leaderboard: ManualScore[];
}

export interface UpdateManualScoreBody {
  tournamentId: string;
  poolMemberId: string;
  /** @nullable */
  r1?: number | null;
  /** @nullable */
  r2?: number | null;
  /** @nullable */
  r3?: number | null;
  /** @nullable */
  r4?: number | null;
  /** @nullable */
  updatedBy?: string | null;
}

export interface CreateTournamentBody {
  name: string;
  year: number;
  espnEventId: string;
  refreshIntervalMinutes?: number;
  password: string;
}

export interface CreatePoolMemberBody {
  name: string;
  password: string;
}

export interface AdminPasswordBody {
  password: string;
}

export interface SavePicksBody {
  tournamentId: string;
  poolMemberId: string;
  golferIds: string[];
  password: string;
}

export interface SavePicksResponse {
  success: boolean;
  message: string;
}

export interface GolferFieldEntry {
  id: string;
  /** @nullable */
  espnId?: string | null;
  name: string;
}

export interface ForceRefreshBody {
  tournamentId: string;
  password: string;
}

export interface RefreshResponse {
  success: boolean;
  message: string;
  /** @nullable */
  lastUpdated?: string | null;
}

export type GetTournamentFieldParams = {
  espnEventId: string;
};