# lib/api-zod/src/generated/types/createTournamentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Golf Pool Scoreboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTournamentBody {
  name: string;
  year: number;
  espnEventId: string;
  refreshIntervalMinutes?: number;
  password: string;
}