# lib/api-zod/src/generated/types/golferFieldEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Golf Pool Scoreboard API
 * OpenAPI spec version: 0.1.0
 */

export interface GolferFieldEntry {
  id: string;
  /** @nullable */
  espnId?: string | null;
  name: string;
}