# lib/api-zod/src/generated/types/manualScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Golf Pool Scoreboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ManualScore {
  id: string;
  tournamentId: string;
  poolMemberId: string;
  poolMemberName: string;
  /** @nullable */
  r1?: number | null;
  /** @nullable */
  r2?: number | null;
  /** @nullable */
  r3?: number | null;
  /** @nullable */
  r4?: number | null;
  /** @nullable */
  toPar?: number | null;
  /** @nullable */
  updatedBy?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}